+++ R/admixcor.R
@@ -12,7 +12,6 @@ admixcor <- function(
                      tol = sqrt( .Machine$double.eps ), # 1e-15
                      nstep_max = 100000,
                      report_freq = 1000,
-                     fix_L = FALSE,
 		     stretch = FALSE,
                      ties_none = FALSE,
 		     tol_stretch = -0.01
@@ -33,7 +32,7 @@ admixcor <- function(
     gamma <- gamma * n / K
     
     # initialize other variables
-    Vars <- initialize( ThetaSR, K, n, Q_type, L_type, fix_L = fix_L )
+    Vars <- initialize( ThetaSR, K, n, Q_type, L_type )
     Q0 <- Vars$Q
     L0 <- Vars$L
     R0 <- Vars$R
@@ -70,12 +69,12 @@ admixcor <- function(
 		Q1 <- stretch_Q( Q1, ties_none = ties_none, tol = tol_stretch )$Q
 		Q1[Q1 < 0] <- 0
 		Q1 <- Q1 / rowSums( Q1 )
-		L1 <- update_L( ThetaSR, Q1, R0, gamma, algorithm = L_algorithm, fix_L = fix_L )
+		L1 <- update_L( ThetaSR, Q1, R0, gamma, algorithm = L_algorithm )
 		R1 <- update_R( ThetaSR, Q1, L1 )
 	}
 	else {
 		R1 <- update_R( ThetaSR, Q0, L0 )
-	        L1 <- update_L( ThetaSR, Q0, R1, gamma, algorithm = L_algorithm, fix_L = fix_L )
+	        L1 <- update_L( ThetaSR, Q0, R1, gamma, algorithm = L_algorithm )
 		Q1 <- update_Q( ThetaSR, L1, R1, delta, I, algorithm = Q_algorithm, vertex_refine = vertex_refine )
 	}
         # calculate step sizes, to assess convergence